requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
volcengine>=1.0.140
openai>=1.30.0
Pillow>=10.0.0
